Changed final death lightning bolt spawn cords
So ~5 was supposed to be high enough to not affect the killer, but I didn't account for the fact that they might be positioned above the victim, or that they might have been knocked upwards a meter by the victim's last strike, which I think happened once.

So now it just spawns underground at y56. If they happen to be there when it occurs... IDK too bad I guess? Might change later.